+++ src/Componenets/App/App.module.scss
@@ -1,5 +1,10 @@
 @import './variables';
 
+body {
+  background: var(--appBackground);
+  color: var(--appTextColor);
+}
+
 .app {
   direction: rtl;
   display: flex;
@@ -10,8 +15,6 @@
   height: 98vh;
   justify-content: space-between;
   margin: 0 auto;
-  background: var(--appBackground);
-  color: var(--appTextColor);
 
   .upper {
     width: 100%;
